# contracts/openapi.json
{"openapi":"3.0.1","info":{"title":"Stripe Gateway API","description":"Fail-closed provider boundary for server-owned checkout creation, signed Stripe webhooks, readiness and owner-scoped Checkout Session expiry.","contact":{"name":"Jobseeker Copilot"},"license":{"name":"MIT"},"version":"2.4.0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/internal/v2/stripe/checkout-sessions/expire":{"post":{"tags":["stripe-lifecycle-controller"],"operationId":"expireOwnedStripeCheckoutSession","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpireOwnedCheckoutSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExpireOwnedCheckoutSessionResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v2/stripe/checkout-sessions":{"post":{"tags":["stripe-v-2-controller"],"operationId":"createOwnedStripeCheckoutSession","parameters":[{"name":"X-Payment-Owner","in":"header","required":true},{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOwnedCheckoutSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateOwnedCheckoutSessionResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/stripe/webhook":{"post":{"tags":["stripe-controller"],"summary":"Receive signed Stripe webhooks","operationId":"webhook","parameters":[{"name":"Stripe-Signature","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/stripe/checkout-sessions":{"post":{"tags":["stripe-controller"],"summary":"Create a Stripe Checkout Session for an AI token bundle","operationId":"createCheckoutSession","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCheckoutSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateCheckoutSessionResponse"}}}}},"security":[{"serviceToken":[]}]}},"/internal/provider-mode":{"get":{"tags":["provider-mode-controller"],"operationId":"providerMode","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/v2/stripe/readiness":{"get":{"tags":["stripe-v-2-controller"],"operationId":"getOwnedStripeReadiness","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/StripeReadinessResponse"}}}}},"security":[{"serviceToken":[]}]}}},"components":{"schemas":{"ExpireOwnedCheckoutSessionRequest":{"required":["orderId","providerSessionId"],"type":"object","properties":{"orderId":{"type":"string","format":"uuid"},"providerSessionId":{"type":"string"}}},"ExpireOwnedCheckoutSessionResponse":{"type":"object","properties":{"orderId":{"type":"string","format":"uuid"},"providerSessionId":{"type":"string"},"status":{"type":"string"},"paymentStatus":{"type":"string"}}},"CreateOwnedCheckoutSessionRequest":{"required":["orderId"],"type":"object","properties":{"orderId":{"type":"string","format":"uuid"}}},"CreateOwnedCheckoutSessionResponse":{"type":"object","properties":{"orderId":{"type":"string","format":"uuid"},"sessionId":{"type":"string"},"url":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"promotionBonusDocumentCredits":{"type":"integer","format":"int32"},"promotionGuaranteed":{"type":"boolean"}}},"CreateCheckoutSessionRequest":{"required":["pricingPlanId","userId"],"type":"object","properties":{"userId":{"type":"string"},"pricingPlanId":{"type":"string"},"tokenAmount":{"minimum":1,"type":"integer","format":"int64"},"priceGbpPence":{"minimum":1,"type":"integer","format":"int64"}}},"CreateCheckoutSessionResponse":{"type":"object","properties":{"sessionId":{"type":"string"},"checkoutUrl":{"type":"string"}}},"StripeReadinessResponse":{"type":"object","properties":{"checkoutAvailable":{"type":"boolean"},"code":{"type":"string","enum":["READY","PAYMENTS_DISABLED","LIVE_RELEASE_NOT_AUTHORISED","STRIPE_CATALOG_NOT_CONFIGURED"]},"mode":{"type":"string","enum":["LIVE","FIXTURE","DISABLED"]}}}},"securitySchemes":{"serviceToken":{"type":"apiKey","description":"Dedicated Payment Gateway service identity.","name":"X-Service-Token","in":"header"}}}}