# src/main/java/com/jobseekercopilot/stripegateway/dto/ExpireOwnedCheckoutSessionResponse.java
package com.jobseekercopilot.stripegateway.dto;

import java.util.UUID;

public record ExpireOwnedCheckoutSessionResponse(
        UUID orderId,
        String providerSessionId,
        String status,
        String paymentStatus) {}